Move `AccountManager` to `:legacy:account` module
