+++ feature/settings/import/src/main/kotlin/app/k9mail/feature/settings/import/ui/OAuthFlowActivity.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import app.k9mail.feature.settings.importing.R
-import com.fsck.k9.preferences.AccountManager
+import app.k9mail.legacy.account.AccountManager
 import com.fsck.k9.ui.base.K9Activity
 import com.google.android.material.textview.MaterialTextView
 import kotlinx.coroutines.flow.Flow
