+++ legacy/account/src/main/java/app/k9mail/legacy/account/AccountRemovedListener.kt
@@ -1,6 +1,4 @@
-package com.fsck.k9
-
-import app.k9mail.legacy.account.Account
+package app.k9mail.legacy.account
 
 fun interface AccountRemovedListener {
     fun onAccountRemoved(account: Account)
