+++ legacy/common/src/main/java/com/fsck/k9/backends/AccountAuthStateStorage.kt
@@ -1,8 +1,8 @@
 package com.fsck.k9.backends
 
 import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.account.AccountManager
 import com.fsck.k9.mail.oauth.AuthStateStorage
-import com.fsck.k9.preferences.AccountManager
 
 class AccountAuthStateStorage(
     private val accountManager: AccountManager,
