+++ legacy/common/src/test/java/com/fsck/k9/account/FakeAccountManager.kt
@@ -1,9 +1,9 @@
 package com.fsck.k9.account
 
 import app.k9mail.legacy.account.Account
-import com.fsck.k9.AccountRemovedListener
-import com.fsck.k9.AccountsChangeListener
-import com.fsck.k9.preferences.AccountManager
+import app.k9mail.legacy.account.AccountManager
+import app.k9mail.legacy.account.AccountRemovedListener
+import app.k9mail.legacy.account.AccountsChangeListener
 import kotlinx.coroutines.flow.Flow
 
 class FakeAccountManager(
