+++ legacy/core/src/main/java/com/fsck/k9/controller/MessageCountsProvider.kt
@@ -1,11 +1,11 @@
 package com.fsck.k9.controller
 
 import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.account.AccountManager
 import app.k9mail.legacy.search.ConditionsTreeNode
 import app.k9mail.legacy.search.LocalSearch
 import app.k9mail.legacy.search.SearchAccount
 import com.fsck.k9.mailstore.MessageStoreManager
-import com.fsck.k9.preferences.AccountManager
 import com.fsck.k9.search.excludeSpecialFolders
 import com.fsck.k9.search.getAccounts
 import com.fsck.k9.search.limitToDisplayableFolders
