+++ legacy/core/src/main/java/com/fsck/k9/controller/NotificationOperations.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9.controller
 
 import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.account.AccountManager
 import app.k9mail.legacy.search.LocalSearch
 import com.fsck.k9.mailstore.MessageStoreManager
 import com.fsck.k9.notification.NotificationController
-import com.fsck.k9.preferences.AccountManager
 import com.fsck.k9.search.isNewMessages
 import com.fsck.k9.search.isSingleFolder
 import com.fsck.k9.search.isUnifiedInbox
