+++ legacy/core/src/main/java/com/fsck/k9/job/K9JobManager.kt
@@ -2,7 +2,7 @@ package com.fsck.k9.job
 
 import androidx.work.WorkManager
 import app.k9mail.legacy.account.Account
-import com.fsck.k9.preferences.AccountManager
+import app.k9mail.legacy.account.AccountManager
 import timber.log.Timber
 
 class K9JobManager(
