+++ legacy/core/src/main/java/com/fsck/k9/preferences/KoinModule.kt
@@ -1,5 +1,6 @@
 package com.fsck.k9.preferences
 
+import app.k9mail.legacy.account.AccountManager
 import com.fsck.k9.Preferences
 import org.koin.core.qualifier.named
 import org.koin.dsl.bind
