+++ legacy/core/src/main/java/com/fsck/k9/search/LocalSearchExtensions.kt
@@ -3,9 +3,9 @@
 package com.fsck.k9.search
 
 import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.account.AccountManager
 import app.k9mail.legacy.search.LocalSearch
 import app.k9mail.legacy.search.SearchAccount
-import com.fsck.k9.preferences.AccountManager
 
 val LocalSearch.isUnifiedInbox: Boolean
     get() = id == SearchAccount.UNIFIED_INBOX
