+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/settings/export/SettingsExportViewModel.kt
@@ -8,10 +8,10 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import app.k9mail.legacy.account.AccountManager
 import com.fsck.k9.helper.SingleLiveEvent
 import com.fsck.k9.helper.measureRealtimeMillis
 import com.fsck.k9.logging.Timber
-import com.fsck.k9.preferences.AccountManager
 import com.fsck.k9.preferences.SettingsExporter
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
